requirements.txt
opencv-python-headless
pytesseract 
pillow 
fastapi 
uvicorn
python-jose[cryptography]
passlib[bcrypt]
sqlalchemy
psycopg2-binary
python-multipart
pydantic[email]
python-dotenv